+++ WHATS_NEW.md
@@ -7,6 +7,7 @@ New in 8.4
 
 Accounting
 - Incomes: add File Attached
+- Expenses: add File Attached
 
 
 New in 8.3